requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pyyaml>=6.0.0
tqdm>=4.66.0
